chore: renames lib/widgets/movement/ to lib/widgets/movements/
